Curve fitting of diffusion data with D=f(c)
This is a finite element based curve fitting code to fit a set of experimental data (xdata & ydata representing a concentration profile). Here, the code uses Fenics as a FEM solver to "solve" the non-linear diffusion PDE and come up with an interpolating function that can be applied to the data to find a best fit.

Output is best-fit parameters of a (a measure of the non-linearity), Do (the constant portion of the diffusivity) and Co (the interface concentration).